fixes #13479 - docker v2 - add model for manifests and update model for tags
This commit contains some initial changes to support the docker v2 model.

This includes the introduction of a new model (manifests) and modifications
to the existing docker tags model.

With docker v2, the docker image model will no longer be needed; however,
removal of it and any associations will be done separately from this commit.