Fixes #13571 - Remote execution integration
Add ability to use the Remote Execution tooling for the content actions
on hosts.

First, we need to use the developer API to define the remote execution features
and templates.

Then, there is ability in UI to jump into the remote execution UI
from the appropriate places (package an errata operations)

It detects, whether the remote execution feature is there.