Fixes #13917: provide error functions to resource get calls.
All resource details pages would 404 but not inform the user if
a link to a nonexistent resource was followed.  This commit provides
a common API Error handler and a function to put the API error into
the Global Messages.

http://projects.theforeman.org/issues/13917
(cherry picked from commit 8d2202608ecde59f23210303451ab3aa4caa57bb)