Fixes #14048 - do not parse iso8601 date twice
The repo export task will covert a date string to iso8601 format. The CV export
task was doing this as well, and then passing the date object to the repo
export task to get re-converted, which resulted in an error.

Instead, just pass the string through directly for later conversion.