fixes #14049: update reindex to include docker content
This commit updates the 'rake katello:reindex' to include indexing
the docker manifests and tags.  The behavior is slightly different
than other content units; however, this is due to the relationship
of tags to a manifest.  For this reason, we will simply perform
the same indexing step that we do within the logic when a repository
is synced...etc. which will recreate, if needed, the manifests
and/org tags.

(cherry picked from commit 753afba4a5a91ea4eb20559ff6ac7c1fa84837a2)