Fixes #14057 - fix broken unit test
The test for uploading files to Pulp was broken in "live" mode but would
succeed when using VCR recordings. This patch adds the needed parameter that
recent Pulp versions check for.

I looked through our unit upload code and it looks like we already set
`unit_key` there, so it's just the test that needs updating.