fixes #14111 - docker repository: update create to require url and upstream name
With these changes, when a docker repo is created, the user must specify
a URL and upstream docker name for the repo. Without it, if a user
created repo without specifying them, they were unable to add them.

Note: making these required for docker repos is Ok, since katello is
no longer able to allow the user to upload content to a repository.