fixes #14186 - dashboard - update usage of ellipsis
Prior to this commit, hovering on one of the katello widgets would
show all content of the widget.  This was due to 'ellipsis' being
applied to the table vs the cell.

This commit will move the 'ellipsis' to those cells where we may
have really long strings (e.g. names).  In addition, it adds the
'table-fixed' class for consistency and to ensure that 'ellipsis'
are used, when needed.

(cherry picked from commit c08cbaafd5e952b4bae4c47bf6cd76e1be950238)