Fixes #14209 - allow for restarting of pulp tasks on resume
The root cause of this is that this commit:

https://github.com/Katello/katello/pull/5567

made it so that resume_external_action no longer threw an exception. Due
to a recent change in dynflow it is needed to throw an exception in the case of
a non-network error.  So this change returns the functionality of checking for
pulp task errors on polling, but overrides poll_external_task_with_rescue to not
treat pulp errors as poll errors.

Conflicts:
	app/lib/actions/pulp/abstract_async_task.rb