fixes #14305 - docker repos in content views should not require url and upstream name
This commit fixes an error that occurs during content view publishing
of views containing docker repositories.  The error was on the validation
that required every docker repository to have a url and upstream_docker_name.
This validation is primarily needed for repositories in the library
that are used for syncing/mirroring that content from the upstream
repository; however, in the case of the content view repositories, they
are simply cloned from repos managed by katello.