Fixes #14388 - Creates RH Atomic OS in seed
When subscription manager sends facts about a content host then the
Operating system of the host is inferred from its distribution name
and version. That inferred OS is attached to the host
However the host validation requires install media/partition tables to
be of the same OS
The facts parser however does not understand what an Atomic OS is.
Thus it ties Atomic host to Redhat 7.2 even for atomic.
This causes host.save! to fail because the right media is NOT attached
to OS.

This commit creates a dedicated  "Redhat Atomic 7.2"  OS in the seed
script  and tell the user if you want to use Atomic use this exact OS.

The facts parser then checks if "distro name has atomic" tie it to this OS

(cherry picked from commit 5926f92b1a19397dd3f57f0d932221279c4d4180)