Fixes #14428 - improves performance with purge empty errata and groups
This improves and streamlines the purge empty errata and purge empty group tasks.
Previously we did not index all of the data until after these steps ran, causing us to
go to pulp for lots of data that could be really slow on heavily loaded systems.

With this change we index prior to these steps and use our database whenever possible.

On a lightly loaded satellite this only saved about 14 seconds for PurgeEmptyErrata
(17s vs 3s) and 10 seconds for PurgeEmptyPackageGroups (11s vs  8s). However for
heavily loaded systems the slowest part of each of the steps no longer exists
(pulling all the package filenames and names for all rpms in the repo).  Some servers
have seen these steps take ~30 mintues or longer due to these calls.