Fixes #14537 - fix `host.organization` references
The patch in #5951 fixed bats, but was not the most correct way of fixing the
issue.

This commit calls `host.save!` before unregistering, which allows use of
`host.organization` again. It also reverts changes to the
`connect_to_smart_proxy` method since we no longer need to use the
content_facet to get the org.