Fixes #14698 - use new ISS default export location
The patch in https://github.com/Katello/puppet-katello/pull/123 creates
a default directory in `/var/lib/pulp/katello-export`.

This Katello patch sets the default to match, which allows users to
perform an export "out of the box" without having to hop to the command
line to create directories.

A new directory under `/var/lib/pulp/` was chosen for the default since
users are already aware that this path may use considerable disk space.
This setting is still settable via web UI and hammer in case someone
wants to do a one-off export to a different location.