Fixes #14802 - Correctly determines if working with Host/Hostgroups page
Prior to this commit the hosts and hostroups pages used the same erbs
They were just identifying what constitutes a host page vs hostgroup
page in the wrong manner.
It used -> @host.nil? to determine this which doesnt work for all
situations.

This commit instead centralizes that logic in a nice helper and uses a
simpler 'check the controller name' approach