Fixes #14916 - restrict https to 'cdn.redhat.com' (#6016)
Katello only ships with a CA cert for `cdn.redhat.com`. However, users
are allowed to put any https URL into the manifest URL page. This
caused TLS sync errors which are not apparent until later when a sync
is kicked off.

Instead, validate the URL to ensure that it is either a `http` URL, or
is `https://cdn.redhat.com`.