Fixes #13762: Don't save organization during environment create (#6025)
Organization save when creating environment could lead to
mismatches that report validation errors to the user on the
organization but not the environment itself. This code previously
ensured the organization was associated to the environment but
testing indicated this was not needed.
(cherry picked from commit 77cdcad709874ca8015db509382abe1ace97d9c4)