fixes #14268 - optimize scoped_search queries when sub_total == 0 (#5972)
There are three queries to the database (more or less) when a request is
passed through Katello::Api::ApiController#scoped_search in the
following order.

Query 1 - Get the total number of rows for a given object type ("total")
Query 2 - Get the total number of rows using a given search string
          ("sub_total")
Query 3 - Get the actual rows on the object type using a given search
          string. ("results")

Query 1 is trivial and essentially instantaneous. Queries 2 and 3 can
take some time though. In the case of an errata index with RHEL7 sync'd,
queries 2 and 3 both take roughly a second bringing the average total
request time over 4 seconds.

This commit adds a small change to #scoped_search to reduce the number
of database queries, namely 2 and 3, where possible. If the total number
of rows for a given object are zero, then we can assume that sub_total
will also be zero and that the result set is going to be an empty array.
It can also be assumed that if sub_total is zero then the result set
will be ``[]``. Knowing these assumptions, we can skip queries 2 and 3
when we know results are going to be empty.
(cherry picked from commit 5dd2cae9777b6796f5f933553856a9765644c8f2)