Fixes #15183 - delay cp content sync till end
of publish and promote tasks.  Prior to this change
when deleting multiple repos during publish & promote (because
the repo is no longer in the Content View), multiple set_content
calls occuring can result in an error.  Since we are syncing
the content ids at the end of this process anyways, they are not
needed.