Fixes #15212 - handle ListenOnCandlepinEvents finish waiting
When the previous tasks haven't finished properly, the listen action is
resumed rather than triggered again, so the triggered_action is nil.
In this case we need to check the status of the task though database.

To reproduce (in production):

1. `service foreman-tasks stop`
2. in foreman-rake console:
   `ForemanTasks.trigger(Actions::Candlepin::ListenOnCandlepinEvents)`
   # it will warn about no executor available but that's expected
3. `service foreman-tasks start` # this way we simulate the resuming
4. `service foreman-tasks stop` while watching the production.log

Without the fix, there will be an error about `nil` not havaing
`finished` method.