Fixes #15233 - add index to `katello_repository_rpms`
The CVV page was calling
`/katello/api/v2/content_view_versions?content_view_id=<id>`, which includes,
among other things, a package count for all repos in each CVV.

This is found via the following query:

```sql
SELECT Count(*)
FROM   "katello_rpms"
       INNER JOIN "katello_repository_rpms"
               ON "katello_repository_rpms"."rpm_id" = "katello_rpms"."id"
WHERE  "katello_repository_rpms"."repository_id" IN (SELECT
       "katello_repositories"."id"
                                                     FROM
       "katello_repositories"
                                                     WHERE
              "katello_repositories"."content_view_version_id" = $1
              AND ( environment_id IS NULL ))
```

This query induced a sequential scan on `katello_repository_rpms` (see
https://explain.depesz.com/s/sMy).

There was an index for `repository_id, rpm_id` for the table, but not just
`repository_id`. This patch adds a migration to create the index.

Times per-query should be reduced from 300+ msec to 20-30 msec. The explain
plan with the index is available at https://explain.depesz.com/s/epd9.

_NOTE_: when testing, you will need to ensure `katello_repository_rpms` has
enough rows for the profiler to prefer the index over a seq scan. I used an
existing DB that was a shade under 4 million rows for the table. If the table
is too small, both explain plans (with and without the index) will look the
same.