Fixes #15177 - Add validation to content view
puppet module to make sure a valid content view with that same uuid
or name/author combination exists. In addition, the user may use the
puppet_module id, which will set the uuid of the
content_view_puppet_module to the same as the puppet_module.

If a content view is specified, the puppet module must belong to a
repository that belongs to the same organization as the content view.

An edge case: If two puppet modules with identical names and authors but
two different organizations are used to validate a content view puppet
module whose content view organization is one of the two organizations,
validation should pass.