fixes #15260 - migrate_content_host: update create_content_facet
This commit updates the 'create_content_facet' so that the caller
can include an argument which will cause it to only associate
lifecycle environment and content view with the facet created.

This is needed to handle the case where the content facet is
created for a host; however, the legacy system will be immediately
unregistered/destroyed.  With this change, when the host
re-registers after the migration is completed, no errors will
be generated (e.g. 404 on delete of pulp consumer).

(cherry picked from commit 0ed008c0f7b3cff848bbe369acfa26ff69ad4181)