Fixes #15162 - Host collections can be removed (#6063)
Allow the last host collection to be removed from an activation key.

This issue is caused by the deep_munge method in Rails 4. We use our API
to GET a record, update a collection of children, then PUT the updated
record to the API in order to update the DB. If the collection of
children is empty, deep munge in conjunction with strong parameters,
translates the empty array into nil, and the collection of children
remains unchanged.

This fix is a workaround and should be removed when upgrading Foretello
to Rails 5. I made a comment containing "Rails 5" so it can be addressed
during the upgrade with a global find.

Add a test for removing all host collections from an activation key.