refs #15297: update puppet repo distributor during upgrade
This commit contains a rake task to be executed as part of the
Katello 3.0 upgrade to ensure that all puppet repositories
(Repository & Content View Puppet Environment) are updated
to include the 'puppet distributor' that was added as part
of Katello 3.0.  In addition, it will publish those repositories
to ensure that the proper content is available on the
filesystem.