Fixes #15389 - errata page is broken
There was a SQL optimization done in PR #6094 which made the dashboard
faster, but interfered with some other callers of the errata model obj.
Basically, the optmized SQL had a `DISTINCT` baked in, so if you called
`uniq` on the results, you'd get an error since you'd have two
`DISTINCT`s.

It turns out that there was an extra inner join, which created a very
large dataset for postgres to work on. For example, 5.5K hosts with 650
errata may create a dataset of upwards of 84M records. This would
exhaust system resources quickly.

This commit removes the extra inner join, and removes the opimization
in #6094 since its no longer needed.

(cherry picked from commit 42502063e7a47176dec428b13aade2c6329438de)