Fixes #15420 - Adjust full_results param (#6126)
While the API advertised a parameter named :full_results, our code was
expecting the param to be named :full_result. This change adds tests to
ensure the :full_results parameter can be used and is actually effective
in our system.

In addition, this changes the API parameter documented from
:full_results to :full_result

Furthermore, this parameter was failing to cast from a JSON string to a
boolean such that 'false', 'true', 0, or 1 would return a true value.
This introduces a proper cast using `Foreman::Cast.to_bool`.