Fixes #14218 - Uses id not uuid, for host link with virtual sub
When attaching a VDC subscription to a physical host, a link appears
for the guest subscription, "Guests of foo.example.com" This link is
broken as it is still using the candlepin id instead of the active
record id. This commit changes the link to use the proper id, and also
changes the JSON returned. All places in the UI that are using a
subscription's host id should be updated.