Fixes #15269: Content host migration breaks loop (#6162)
The migration of content hosts to hosts has a loop over systems
as it updates or destroys them. In the case of systems without
a Candlepin fact hostname, the loop was being broken instead of
moving to the next system in line. This led to systems never
being properly migrated that came after a system without a hostname
(cherry picked from commit b5f07391221ca330bade21d1d1614e1e4e2ee4f4)