Refs #14372 - global status not updating correctly on host w/ CV (#6166)
When creating a host with a content view and lifecycle environment,
the global status will not be correct as the errata status will be
marked as relevant when it is not, since the host is still not
registered with pulp. This will check for a uuid, which is only
present if the host is registered with pulp.