Refs #14372 - global status not updating correctly on host w/ CV (#6166)
When creating a host with a content view and lifecycle environment,
the global status will not be correct as the errata status will be
marked as relevant when it is not, since the host is still not
registered with pulp. This will check for a uuid, which is only
present if the host is registered with pulp.
(cherry picked from commit bab5613eb58a352f7e5e7f7133536ccb1af1b553)

Conflicts:
	app/models/katello/errata_status.rb