Refs #12556 - remove System from views & controllers (#6156)
This attempts to remove any references to the System
model from all controllers and views.  The only exception to
this is the CandlepinProxies controller which relies on creation and
updating of the models during sub-man actions.  These will be removed
along with the model.

Some areas to test more heavily are:
  * Content View Version Removal
  * Various Content Host pages

A followup commit will remove the model and tests.