fixes #15678 - move validation to content view model
This commit moves one of the validations performed by the
content_view_component to the content_view model.

This is to address an error that was observed on older
rails (4.1) where attempting to create a content view
with components would result an error.

The following is an example of the error when using
hammer-cli:

hammer> content-view create --name test-ccv
--organization-label Default_Organization
--composite --component-ids 2

Could not create the content view:
  undefined method `composite?' for nil:NilClass