Fixes #15824 - enabling a repository should fail on pulp error (#6207)
When enabling a repository, if there is a pulp error, the task will
be paused after creating the repository in the database, though
it is not created in pulp (since it errored). Moving this to create
in plan phase of Katello::Repository::Create ensures that the task
will stop and not complete on a pulp error, leaving the repository
disabled.