fixes #15531 - repo: bypass AR validation when deleting units
When a repository is destroyed, many of the related objects
are also destroyed.  In the case of AR objects that are related
to 'pulp database units' (e.g. rpm, erratum, ...etc), a
'dependent destroy' was performed.  This triggers AR
validations to be performed and cause the deletion to
take a very long time (possibly minutes vs seconds).
This change will bypass AR for the deletion of anything that
is associated with a pulp database unit.

(cherry picked from commit cc350d92d3684b7f02110ec871900ddc70199a35)