Fixes #16177 - Unassociate pulp units on capsule sync (#6270)
When syncing a capsule, if there is a change in versions of a unit
in a repo (like puppet module versions), pulp will try to sync
both and publish, which results in an error. This can be avoided
by first unassociating all units from non-yum repos and then
kicking off the sync.