Fixes #16713: do not show subscription info if no org is selected (#6357)
Previously, the dashboard would iterate over each organization the user
had access to and sum up the subscription info. If the user had a large
number of orgs, this could make the dashboard extremely slow, since
each organization required two calls to candlepin to find the
subscription counts.

Instead, display a message stating that an organization needs to be
selected in order to show subscription info. It is unlikely a user
would want to see a "roll-up" view of subs across orgs, since
subscriptions are tightly coupled to a particular org.