Fixes #16733: raise useful error when exporting non-yum repo
ISS currently only supports exporting yum repositories. However, this
was not enforced by the controller, and if you attempted to export a
repo that was not yum, you would get a baffling error message in the
action.

This patch checks that the repo is the correct type before exporting,
and raises a 400 error if not.