Fixes #16893 - rely on DB to determine errata type in collection
Previously, the errata checks in host collections would iterate over
all errata in memory to determine if any errata were critical. If there
were a lot of errata to iterate over (say, over 500K), this could take
nearly a minute and force passenger to consume lots of CPU.

This patch alters this behavior so a DB query is used, times appear to
be sub-second now.