fixes #17115 - validate ip from RHSM
On older versions of RHSM, ip values can be sent as a
string 'Unknown'. Foreman historically saves hosts
facts without validation, which means the bogus
values get saved. When remote execution then tries to
use that interface's IP it hits an error.

This adds some validation to the RHSM fact parser, and
a migration to fix the bad values we may have created.