Fixes #17266 - Fix tests that depend on CVE 2016-7078
A small number of tests in the Katello codebase depended on regular
users being able to see objects without organization/location. This is
now fixed in core through a CVE (users shouldn't view stuff they're not
scoped to see), so in order for Jenkins to pass, we need to make Katello
tests pass with it too.