Fixes #17388: fix display of new repository page.
The new repository page wasn't using the correct ui-view and thus
wouldn't show up when visited.  This commit changes the ui-router
configuration to make the page display as well as renaming the new
repository page for consistency and fixing a display issue on that page
where the breadcrumbs were being shown twice.

http://projects.theforeman.org/issues/17388