Fixes #17512 - index content in a paged manner
This does two major things:

 1) Centralizes fetching and indexing of content to central places
(pulp_database_unit.rb & pulp_content_unit.rb). This should help making
content units pluggable in the future easier, as all the code is based
around the unit DB class or service class.

 2) Change the indexing so that we only keep a page (default is 500) of
units in memory. Previously we would fetch all units in pages, but then
take the entire list and index that. With this change we actually index
just the page while we have it, and then throw it away to fetch another
page and process that.