fixes #17524 - capsule - update repo creation during repo sync
Currently, when a user syncs a repository (in library) that will
be synced to a capsule, the logic will create all missing
repositories on the capsule.  This means that syncing one
repository on a brand new capsule could result in tens or
hundreds of repositories being created.

This commit changes the behavior such that only the repository
being synced will be created.   That said, if a user performs
a full sync of the capsule (e.g. from cli or ui via
Infrastructure -> Capsule (or Smart Proxies)), all repositories
will be created.