Fixes #17564 - use threading to improve RH repos page
This commit spawns threads in order to speed up the
Red Hat Repositories page for fetching listing files.

Given a directory structure such as:

/path/$RELEASEVER/$ARCH

we have to make 1 + $RELEASEVER number of requests. On
a high latency connection this can take some time per request.
With this change we put each request in a thread and so are able
speed up the total time to be closer to the same amount of time
as two requests.

In addition, this refactors the code removing features of the
cdn_var_substitutor that are no longer needed and making the code
much more readable and better tested.