Fixes #17622: Display need for virt-who in the UI
Uses the virt_limit property of a pool and if its a Red Hat
subscription to determine if a user needs virt-who. This is displayed
on the subscription list, and subscription details. As well, if
a manifest includes subscriptions that require virt-who an info
alert is displayed to point them to the Red Hat virtual instance
guide.