Fixes #17719 - katello:check_ping should load environment
If the task doesn't load the environment and is ran through
`foreman-rake katello:check_ping` we get `uninitialized constant User`.
Today I wanted to use it on its own and there was no way to do it. It's
listed on the list of tasks that one sees through `rake -T` so I can see
more people wanting to try it out.