Fixes #17610 - Corrects multi repo cv publishes
Commit e5586b7e6bfd07455997fdb1d7651650f9a83154 broke publishing of
non-composite content views with multiple repositories. This is because
the changes were meant more for composite cv's and assumed library
instances of repositories to be published to be non nil.
In short you could not add a yum repo and repo of another type and
publish em together.
This commit fixes that issue by treating repos with no library
instances differently from  ones with library instances (which tend to
be the component cv repos.)