Fixes #18059 - Rendering RABL for host/main fails
Currently, Katello does not add its views to the Rabl paths. For this
reason, when I try to run a playbook with foreman_ansible, I get an
error that  katello/api/v2/content_facet/base_with_root isn't in its
view paths

foreman_ansible InventoryCreator is trying to render the
api/v2/hosts/main using Rabl, however it fails to do so as the
content_facet isn't available to it.

Solution: add an initializer to configure view paths for Rabl in the
Katello engine