Fixes #18122 - move rhsm fact update to run phase
During fact importing, RhsmFactName objects are created
in large quantity. Because first_or_create is very susceptible
to race conditions, we rescue on RecordNotUnique within
RhsmFactImporter#add_fact_name.  However if duplicate record
exception is thrown while in a transaction the entire
transaction is aborted.  The run phase does not use a transaction
so the problem should not occur there.