Fixes #18166 - Add required helpers to Sync controller
Notifications will not be able to run without that helper on this page,
as mount_react_component will not be available. Also the
notifications_recipients_path will not be available if the
proper url helpers are not included in the controller. The
reason this is needed now is because the user dropdown in
the top-right corner is calling those helpers on #18010