Fixes #18226 - Allow default Location to be removed
The default location provided a location for new Puppet content
publishes and for new hosts coming through subscription manager. This PR
adds a setting for these two things, which would allow an user to change
the location they want for each of these things separately. It still
protects that the setting must always have a value