Fixes #18140: replace search on CH errata modal with filter.
The search on the content host errata modal does not work because of the
requirement of using HostBulkAction.installableErrata() to display the
errata for the selected content hosts.  Instead of using search we
should use a filter. Also fix issue with the input not retaining focus
by using the stop event propagation directive workaround.

http://projects.theforeman.org/issues/18140