Fixes #18036 - Correctly sets  sync plan schedules
This commit fixes a bug where a disabled sync plan caused an ISE when
a new repository was being created in the product associated to the
sync plan. The issue occured when there was an attempt to calculate the
"next_sync_date" even if the sync plan was disabled.
This commit fixes that issue by making sure that the next sync date
would not get calculated if the plan is disabled.