Fixes #17949 - adjust generation of pulp_id
This changes how we generate a pulp_id.  Now
for all library instances, we generate a random uuid
and then for each clone we construct a combination of
org id, env label, cv label, and the library instance's uuid.

This also seperates the pulp_id and the docker repository name in pulp
allowing for these to be different