Fixes #18032 - Skip pulp metadata regeneration on CVV promote
If nothing has changed, we have no reason to regenerate metadata
in the pulp repo. The way things currently are we regenerate the
metadata (even if nothing has changed in the repo) and it is
slowing down Smart Proxy Content syncs. Not regenerating the metadata
will take full advantage of certain optimizations in pulp making
these Smart Proxy syncs much faster.

This commit adds an option to force the metadata regeneration,
it is a checkbox called "Force Metadata Regeneration" on
Content View publish and promote.