Fixes #18459: fix incremental update select content host display
The next and cancel buttons for incremental update were being displayed
below the table and if there were a lot of content hosts the user would
have to scroll through the entire list of content hosts (due to infinite
scrolling) in order to get to the buttons.  This commit temporarily
moves the next/cancel buttons to the top of the table until this entire
workflow can be replaced by the angular-patternfly wizard.

http://projects.theforeman.org/issues/18459