Fixes #18655 - Only index new rpms on repository sync
We index all rpms after a repository sync, which is unnecessary
for the already present rpms (in katello db). We should only
index the rpms that are not present in katello's db. This
new logic will say "What rpms are in pulp that are not in
our db for this repository?" and index those.