Fixes #18781 - CSV export for products table
Since we are providing the ability to export to CSV on
projects.theforeman.org/issues/18670 , I think products is a table that
could easily use this feature.
This ticket is meant just as a way to introduce the feature in Katello
(reusing the infrastructure in Foreman), and the aim is to extend it
over to more tables.

This would require writing extending the table Angular directive to know
these APIs exist.