Fixes #18416 - Weight capsule sync task
SyncCapsule task will take the longest by far during a capsule
sync so we should weight it heavily.

Once the SyncCapsule task initiates the pulp tasks, the
progress_weight of that step jumps to 0.5 (half done).
This is resulting in the progress bar jumping to over
half done very shortly after the sync is executed, even
if it will take hours more. This commit will modify each
SyncCapsule step to have a 0.1 progress weight until it
finishes. As the SyncCapsule steps finish, the progress
bar will move incrementally at a more even pace.