Fixes #19061 - kick off async fact update
Previously, the POST call to update facts waited on a synchronous
task. This meant that if there was a backlog of tasks, passenger
workers would start to fill up, causing katello to fall over.

Instead, just kick off an async task to update facts. If the task
fails, users may not know without checking the task status, but it
should be OK to do this given the recurring nature of fact updates.