Refs #19189 - Check if adapter includes sqlite
Turns out the adapter is SQlite which downcases to sqlite and does
not match the name of the adapter that one would specify in
the database.yml. This switches to check for include to match sqlite
in case Rails for some reason changes this name.

(cherry picked from commit 35de7e892031410d4894e7e2378f9ebe0a38349e)