Fixes #19076 - use new runcible call for applicability
The previous consumer regen call would 'pin' to a single worker[0].

Instead, use an API call to Pulp that can be spread across multiple
workers. Note that this requires
https://github.com/Katello/runcible/pull/189 to work.

I had to rename `name` to `consumer_name` to work around a tricky
issue. There was an interaction between minitest and VCR where it
would see a var named `name` in a test class, get confused, and pass
all the tests.

[0] https://pulp.plan.io/issues/1173#note-12