Fixes #19302 - Notification for Pulp low disk space
If some proxy is running out of disk space, this notification will warn
the user about it. This will check all proxies with the Pulp feature for
storage on a cron and create notifications according to that. IF the
notification already exists, then it will not create a new one.