fixes #19323 - errata/package applicablility should filter on org
The Content -> Errata and Content -> Packages are not taking
in to consideration the organization when determining if the
content (Errata or Packages) are applicable/upgradeable/installable
for hosts.  In other words, it is looking at all hosts
across all organizations in the current logic.

(cherry picked from commit 6d7d85f04e321d6947b18f63b6932e5f0ad94af1)