Fixes #19468 - Displays schema version for Docker
Docker V2 has 2 different schema types Schema 1 and 2.
Pulp 2.13 introduces syncing manifests of the 2 different schemas
from a docker registry repo.
This means the same tag in the same repo might point to 2 different
manifests.
Related to this this commit updated the following pages to show the
Schema Version

1) Repository -> Manage Docker Manifests
2) Repository -> Manade Docker Tags
3) Content -> Docker Tags (search page.)

In addition a new DockerMetaTag model got introduced. The role of this
model is to maintain association between the schema v1 version of the
tag and the v2 version of the tag.