Fixes #18845 - Reduce n+1 queries for hosts
this reduces the number of N+1 queries considerably
via three methods:

1) by fetching errata counts using a group queries
to fetch counts for all types in one query. Note this
also affects content view versions and repos as well.
2) By using the cached status when fetching the
errata status rather than re-computing it at read time
3) By not including :applicable_errata in the 'include'
optimization, as its not needed

This change takes the number of queries for 3 systems (2 with errata)
from 15 queries to 2, .58ms to .6ms.

Further improvements could be made to remove all queries, but this
would require caching the data on the host object (or facet), as well
as CVV and repos.  I'm not quite ready to do that yet.