Fixes #19741 - Undefined method split for nil on tests
Due to a bug on minitest that is being currently fixed, our tests are
broken when they run as the "test_develop_pr_katello" job in Jenkins.

The fix is published as https://github.com/seattlerb/minitest/pull/696
however there is no version of minitest with it yet. A possible
workaround is to prevent these assertions like this PR does.